JAVA-251 - Support new 1.8 MapReduce options, as well as inline results output.

+++ src/main/com/mongodb/MapReduceOutput.java
@@ -11,8 +11,15 @@ public class MapReduceOutput {
     MapReduceOutput( DBCollection from , BasicDBObject raw ){
         _raw = raw;
 
-        _collname = raw.getString( "result" );
-        _coll = from._db.getCollection( _collname );
+        if ( raw.containsKey( "results" ) ) {
+            _coll = null;
+            _collname = "<<INLINE>>";
+            _resultSet = (Iterable<DBObject>) raw.get( "results" );
+        } else {
+            _collname = raw.getString( "result" );
+            _coll = from._db.getCollection( _collname );
+            _resultSet = _coll.find();
+        }
         _counts = (BasicDBObject)raw.get( "counts" );
     }
 
@@ -20,19 +27,21 @@ public class MapReduceOutput {
      * returns a cursor to the results of the operation
      * @return
      */
-    public DBCursor results(){
-        return _coll.find();
+    public Iterable<DBObject> results(){
+        return _resultSet;
     }
 
     /**
      * drops the collection that holds the results
      */
     public void drop(){
-        _coll.drop();
+        if ( _coll != null)
+            _coll.drop();
     }
     
     /**
      * gets the collection that holds the results
+     * (Will return null if results are Inline)
      * @return
      */
     public DBCollection getOutputCollection(){
@@ -50,6 +59,7 @@ public class MapReduceOutput {
     final BasicDBObject _raw;
 
     final String _collname;
+    final Iterable<DBObject> _resultSet;
     final DBCollection _coll;
     final BasicDBObject _counts;
 }
